Conflicts:
	build/CEGUI.log
	build/config.cfg

+++ meteor-falls-src/src/Engine/GameEngine/Unites/typedefs.h
@@ -1,6 +1,6 @@
 #ifndef TYPEDEFS_UNITE_H_INCLUDED
 #define TYPEDEFS_UNITE_H_INCLUDED
 
-typedef unsigned short UnitId;
+typedef unsigned int UnitId;
 
 #endif // TYPEDEFS_H_INCLUDED
